Release 0.6.0

+++ grader_service/handlers/submissions.py
@@ -5,6 +5,7 @@
 # LICENSE file in the root directory of this source tree.
 # grader_s/grader_s/handlers
 import json
+import logging
 import shutil
 import time
 
@@ -48,7 +49,7 @@ def remove_points_from_submission(submissions):
 
 
 @register_handler(
-    path=r'\/lectures\/(?P<lecture_id>\d*)\/assignments' +
+    path=r'\/api\/lectures\/(?P<lecture_id>\d*)\/assignments' +
          r'\/(?P<assignment_id>\d*)\/submissions\/?',
     version_specifier=VersionSpecifier.ALL,
 )
@@ -253,7 +254,11 @@ class SubmissionHandler(GraderBaseHandler):
                             reason="Cannot submit completed assignment!")
         if role.role == Scope.student and assignment.status != "released":
             raise HTTPError(HTTPStatus.NOT_FOUND)
-        submission_ts = datetime.datetime.utcnow()
+        # set utc time
+        submission_ts = datetime.datetime.now(datetime.timezone.utc)
+        # use implicit utc time to compare with database objects
+        submission_ts = submission_ts.replace(tzinfo=None)
+        
 
         score_scaling = 1.0
         if assignment.duedate is not None:
@@ -319,7 +324,7 @@ class SubmissionHandler(GraderBaseHandler):
                 grading_chain = chain(
                     autograde_task.si(lecture_id, assignment_id, submission.id),
                     generate_feedback_task.si(lecture_id, assignment_id, submission.id),
-                    lti_sync_task.si(lecture_id, assignment_id, submission.id, sync_on_feedback=True)
+                    lti_sync_task.si(lecture_id, assignment_id, submission.id, feedback_sync=True)
                 )
             else:
                 grading_chain = chain(autograde_task.si(lecture_id, assignment_id, submission.id))
@@ -332,6 +337,7 @@ class SubmissionHandler(GraderBaseHandler):
     @staticmethod
     def calculate_late_submission_scaling(assignment, submission_ts, role: Role) -> float:
         assignment_settings = AssignmentSettingsModel.from_dict(json.loads(assignment.settings))
+        # make duedate aware
         if assignment_settings.late_submission and len(assignment_settings.late_submission) > 0:
             scaling = 0.0
             if submission_ts <= assignment.duedate:
@@ -357,7 +363,7 @@ class SubmissionHandler(GraderBaseHandler):
 
 
 @register_handler(
-    path=r'\/lectures\/(?P<lecture_id>\d*)\/assignments\/' +
+    path=r'\/api\/lectures\/(?P<lecture_id>\d*)\/assignments\/' +
          r'(?P<assignment_id>\d*)\/submissions\/(?P<submission_id>\d*)\/?',
     version_specifier=VersionSpecifier.ALL,
 )
@@ -434,31 +440,43 @@ class SubmissionObjectHandler(GraderBaseHandler):
         :type assignment_id: int
         :param submission_id: id of the submission
         :type submission_id: int
-        :raises HTTPError: throws err if assignment was not found or deleted
+        :raises HTTPError: if submission can't be deleted due to it having feedback
         """
         lecture_id, assignment_id, submission_id = parse_ids(lecture_id, assignment_id, submission_id)
         self.validate_parameters()
         submission = self.get_submission(lecture_id, assignment_id, submission_id)
 
-        previously_deleted = (
-            self.session.query(Submission)
-            .filter(
-                Submission.id == submission_id,
-                Submission.assignid == assignment_id,
-                Submission.deleted == DeleteState.deleted,
-            )
-            .one_or_none()
-        )
-        if previously_deleted is not None:
-            self.session.delete(previously_deleted)
-            self.session.commit()
+        if submission is not None:
+            if submission.feedback_status is not "not_generated":
+                raise HTTPError(HTTPStatus.FORBIDDEN, reason="Only submissions without feedback can be deleted.")
+            # if assignment has deadline
+            if submission.assignment.duedate:
+                # if assignment's deadline has passed
+                if submission.assignment.duedate < datetime.datetime.now().replace(tzinfo=None):
+                    raise HTTPError(HTTPStatus.FORBIDDEN, reason="Submission can't be deleted, due date of assigment "
+                                                                 "has passed.")
+            else:
+                previously_deleted = (
+                    self.session.query(Submission)
+                    .filter(
+                        Submission.id == submission_id,
+                        Submission.assignid == assignment_id,
+                        Submission.deleted == DeleteState.deleted,
+                    )
+                    .one_or_none()
+                )
+                if previously_deleted is not None:
+                    self.session.delete(previously_deleted)
+                    self.session.commit()
 
-        submission.deleted = DeleteState.deleted
-        self.session.commit()
-
+                submission.deleted = DeleteState.deleted
+                self.session.commit()
+        else:
+            raise HTTPError(HTTPStatus.NOT_FOUND,
+                            reason="Submission to delete not found.")
 
 @register_handler(
-    path=r'\/lectures\/(?P<lecture_id>\d*)\/assignments\/' +
+    path=r'\/api\/lectures\/(?P<lecture_id>\d*)\/assignments\/' +
          r'(?P<assignment_id>\d*)\/submissions\/(?P<submission_id>\d*)\/logs\/?',
     version_specifier=VersionSpecifier.ALL,
 )
@@ -488,7 +506,7 @@ class SubmissionLogsHandler(GraderBaseHandler):
 
 
 @register_handler(
-    path=r'\/lectures\/(?P<lecture_id>\d*)\/assignments\/' +
+    path=r'\/api\/lectures\/(?P<lecture_id>\d*)\/assignments\/' +
          r'(?P<assignment_id>\d*)\/submissions\/(?P<submission_id>\d*)\/' +
          r'properties\/?',
     version_specifier=VersionSpecifier.ALL,
@@ -516,8 +534,7 @@ class SubmissionPropertiesHandler(GraderBaseHandler):
         lecture_id, assignment_id, submission_id = parse_ids(
             lecture_id, assignment_id, submission_id
         )
-        properties = self.session.query(SubmissionProperties).get(
-            submission_id)
+        properties = self.session.get(SubmissionProperties, submission_id)
         if properties is not None and properties.properties is not None:
             # delete source cells from properties if user is student
             if self.get_role(lecture_id).role == Scope.student:
@@ -591,7 +608,7 @@ class SubmissionPropertiesHandler(GraderBaseHandler):
 
 
 @register_handler(
-    path=r'\/lectures\/(?P<lecture_id>\d*)\/assignments\/' +
+    path=r'\/api\/lectures\/(?P<lecture_id>\d*)\/assignments\/' +
          r'(?P<assignment_id>\d*)\/submissions\/(?P<submission_id>\d*)\/edit\/?',
     version_specifier=VersionSpecifier.ALL,
 )
@@ -721,7 +738,7 @@ class SubmissionEditHandler(GraderBaseHandler):
 
 
 @register_handler(
-    path=r"\/lectures\/(?P<lecture_id>\d*)\/assignments\/(?P<assignment_id>\d*)\/submissions\/lti\/?",
+    path=r"\/api\/lectures\/(?P<lecture_id>\d*)\/assignments\/(?P<assignment_id>\d*)\/submissions\/lti\/?",
     version_specifier=VersionSpecifier.ALL,
 )
 class LtiSyncHandler(GraderBaseHandler):
@@ -730,17 +747,17 @@ class LtiSyncHandler(GraderBaseHandler):
     @authorize([Scope.instructor])
     async def get(self, lecture_id: int, assignment_id: int):
         # apply task synchronously without adding to queue
-        results = lti_sync_task.apply((lecture_id, assignment_id, None, False))
+        results = lti_sync_task.delay(lecture_id, assignment_id, None, False)
+        results = results.get(timeout=120)
         if results is None:
-            self.log.info("Skipping LTI plugin as it is not enabled")
-            self.write_error(HTTPStatus.CONFLICT, reason="LTI Plugin is not enabled")
+            raise HTTPError(HTTPStatus.NOT_FOUND, reason="Did not find syncable submissions.")
         else:
             self.write_json(results)
 
 
 
 @register_handler(
-    path=r"\/lectures\/(?P<lecture_id>\d*)\/assignments\/(?P<assignment_id>\d*)\/submissions\/count\/?"
+    path=r"\/api\/lectures\/(?P<lecture_id>\d*)\/assignments\/(?P<assignment_id>\d*)\/submissions\/count\/?"
 )
 class SubmissionCountHandler(GraderBaseHandler):
     """
@@ -749,7 +766,8 @@ class SubmissionCountHandler(GraderBaseHandler):
 
     @authorize([Scope.student, Scope.tutor, Scope.instructor])
     async def get(self, lecture_id: int, assignment_id: int):
-        """ Returns the count of submissions made by the student for an assignment.
+        """ Returns the count of submissions made by the student for an assignment, no matter if submission
+         was deleted or not.
 
         :param lecture_id: id of the lecture
         :type lecture_id: int
